Literate Documentation For R.

Literate Documentation For R.

Given a local package directory, turn the package's R files into a literately documented website.

rocco(directory, output_dir = tempdir(), browse = interactive(),
  rocco = TRUE, staticdocs = TRUE, gh_pages = FALSE)

Arguments

directory
character. The package directory to document.

output_dir
character. The directory to output the static HTML site. By default, an extemporaneously-generated temporary directory.

browse
logical. Whether or not to launch the documentation immediately for browsing. This will be set to interactive(), that is, TRUE if the R session is running interactive and FALSE otherwise.

rocco
logical. Whether or not to create rocco docs. Defaults to TRUE.

staticdocs
logical. Whether or not to create staticdocs. Staticdocs are from Hadley's staticdocs package.

gh_pages
logical. If set to true, rocco docs will be served on your gh-pages branch.

Value

TRUE, plus additional side effects are the creation of the documentation in the output_dir and the launching of the browser if browse = TRUE.

Examples

not_run({
    rocco("/path/to/package") # Will create a temporary directory and
      # display literate documentation for everything in the R directory
      # of the package at /path/to/package.

    # The below will simply create a static HTML site without opening it.
    rocco("/path/to/package", output_dir = "/my/html/dir", browse = FALSE)
})